feat(runWindows) - refactoring files
